chore: Update .gitignore and netlify.toml, and add lazy loading for images
This commit updates the .gitignore file to include the "public" directory, ensuring that it is ignored by Git. It also updates the netlify.toml file to change the "publish" value to "public" for the build command. Additionally, it adds lazy loading for images in the layout components of the 2023 and 2024 apps, improving the page load performance.

Fixes #129